requirements.txt
python-telegram-bot==20.7
transformers==4.36.2
torch==2.1.2
torchaudio==2.1.2
librosa==0.10.0
numpy==1.24.3
soundfile==0.12.1
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
